refactor logic for setting azure span tags in serverless mini agent
