+++ trace-mini-agent/src/trace_processor.rs
@@ -9,15 +9,15 @@ use log::info;
 use tokio::sync::mpsc::Sender;
 
 use datadog_trace_obfuscation::obfuscate::obfuscate_span;
-use datadog_trace_utils::trace_utils;
+use datadog_trace_utils::trace_utils::{self};
 use datadog_trace_utils::trace_utils::SendData;
-
-use ddcommon::azure_app_services;
 
 use crate::{
     config::Config,
     http_utils::{self, log_and_create_http_response},
 };
+
+const MINI_AGENT_VERSION: &str = env!("CARGO_PKG_VERSION");
 
 #[async_trait]
 pub trait TraceProcessor {
@@ -80,35 +80,7 @@ impl TraceProcessor for ServerlessTraceProcessor {
                 );
                 for span in chunk.spans.iter_mut() {
                     trace_utils::enrich_span_with_mini_agent_metadata(span, &mini_agent_metadata);
-                    match azure_app_services::get_function_metadata() {
-                        Some(aas_metadata) => {
-                            let aas_tags = [
-                                ("aas.resource.id", aas_metadata.get_resource_id()),
-                                (
-                                    "aas.environment.extension_version",
-                                    aas_metadata.get_extension_version(),
-                                ),
-                                (
-                                    "aas.environment.instance_id",
-                                    aas_metadata.get_instance_id(),
-                                ),
-                                (
-                                    "aas.environment.instance_name",
-                                    aas_metadata.get_instance_name(),
-                                ),
-                                ("aas.environment.os", aas_metadata.get_operating_system()),
-                                ("aas.resource.group", aas_metadata.get_resource_group()),
-                                ("aas.site.name", aas_metadata.get_site_name()),
-                                ("aas.site.kind", aas_metadata.get_site_kind()),
-                                ("aas.site.type", aas_metadata.get_site_type()),
-                                ("aas.subscription.id", aas_metadata.get_subscription_id()),
-                            ];
-                            aas_tags.into_iter().for_each(|(name, value)| {
-                                span.meta.insert(name.to_string(), value.to_string());
-                            });
-                        }
-                        None => (),
-                    }
+                    trace_utils::enrich_span_with_azure_metadata(span, MINI_AGENT_VERSION);
                     obfuscate_span(span, &config.obfuscation_config);
                 }
             },
